pyproject.toml
[tool.poetry]
name = "pipen-lock"
version = "0.8.1"
description = "Process lock for pipen to prevent multiple runs at the same time"
authors = ["pwwang <pwwang@pwwang.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pipen-annotate = { version = "^0.15.0a1", allow-prereleases = true }
filelock = "^3"

[tool.poetry.plugins.pipen]
lock = "pipen_lock:pipen_lock_plugin"

[tool.poetry.build]
generate-setup-file = true

[tool.poetry.group.examples.dependencies]
python-dotenv = "^1.0"
google-cloud-storage = "^3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
allow_redefinition = true
disable_error_code = ["attr-defined", "no-redef", "union-attr"]
show_error_codes = true
strict_optional = false
